[에러수정] 아카이브 기록하기 (#106)
* Fix: 아카이브 기록하기에서 별점 기본 5개, 등록시 알람팝업 하나, 기록후 마이아카이브로 넘어가기

* Fix: 기본 별점 5점시 변경가능하게

* Fix: useRecoilValue로 수정

* Fix: 오타수정
